+++ RELEASENOTES.md
@@ -7,6 +7,7 @@
 *   Avoid throwing an exception while parsing fragmented MP4 default sample
     values where the most-significant bit is set
     ([#7207](https://github.com/google/ExoPlayer/issues/7207)).
+*   Fix `AdsMediaSource` child `MediaSource`s not being released.
 *   DASH:
     *   Merge trick play adaptation sets (i.e., adaptation sets marked with
         `http://dashif.org/guidelines/trickmode`) into the same `TrackGroup` as